// personal-knowledge-base/assets/cloudflare-gateway-template/mcp-configuration.d.ts
/* eslint-disable */
// Generated by Wrangler by running `wrangler types --config=wrangler.mcp.jsonc --include-runtime=false --strict-vars=false --env-interface=McpBindings mcp-configuration.d.ts` (hash: 929f1c4e57724958b727d847c22051b2)
interface __BaseEnv_McpBindings {
	SYNC_STATE: KVNamespace;
	OAUTH_KV: KVNamespace;
	GITHUB_OWNER: string;
	GITHUB_REPOSITORY: string;
	MCP_ALLOWED_LOGIN: string;
	MCP_ORIGIN: string;
	GITHUB_MCP_TOKEN: string;
	GITHUB_OAUTH_CLIENT_ID: string;
	GITHUB_OAUTH_CLIENT_SECRET: string;
	ADMIN_TOKEN: string;
}
declare namespace Cloudflare {
	interface GlobalProps {
		mainModule: typeof import("./src/mcp-index");
	}
	interface Env extends __BaseEnv_McpBindings {}
}
interface McpBindings extends __BaseEnv_McpBindings {}
type StringifyValues<EnvType extends Record<string, unknown>> = {
	[Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
};
declare namespace NodeJS {
	interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, "GITHUB_OWNER" | "GITHUB_REPOSITORY" | "MCP_ALLOWED_LOGIN" | "MCP_ORIGIN" | "GITHUB_MCP_TOKEN" | "GITHUB_OAUTH_CLIENT_ID" | "GITHUB_OAUTH_CLIENT_SECRET" | "ADMIN_TOKEN">> {}
}

export type { McpBindings };